fix: use asWebviewUri; move logic into USS action
