+++ packages/zowe-explorer/src/uss/init.ts
@@ -13,7 +13,7 @@ import * as globals from "../globals";
 import * as vscode from "vscode";
 import * as ussActions from "./actions";
 import * as refreshActions from "../shared/refresh";
-import { IZoweUSSTreeNode, IZoweTreeNode, IZoweTree, Vite, Gui } from "@zowe/zowe-explorer-api";
+import { IZoweUSSTreeNode, IZoweTreeNode, IZoweTree } from "@zowe/zowe-explorer-api";
 import { Profiles } from "../Profiles";
 import * as contextuals from "../shared/context";
 import { getSelectedNodeList } from "../shared/utils";
@@ -21,8 +21,6 @@ import { createUSSTree } from "./USSTree";
 import { initSubscribers } from "../shared/init";
 import { ZoweLogger } from "../utils/LoggerUtils";
 import { TreeViewUtils } from "../utils/TreeViewUtils";
-import { Utilities } from "@zowe/cli";
-import { permStringToOctal } from "@zowe/zowe-explorer-api/src/utils";
 
 export async function initUSSProvider(context: vscode.ExtensionContext): Promise<IZoweTree<IZoweUSSTreeNode>> {
     ZoweLogger.trace("init.initUSSProvider called.");
@@ -144,83 +142,9 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
         vscode.commands.registerCommand("zowe.uss.editFile", (node: IZoweUSSTreeNode): void => node.openUSS(false, false, ussFileProvider))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.editAttributes", (node: IZoweUSSTreeNode) => {
-            const editView = new Vite.View(
-                `Edit Attributes: ${node.label as string}`,
-                "edit-attributes",
-                context,
-                // eslint-disable-next-line @typescript-eslint/no-misused-promises
-                async (message: any) => {
-                    switch (message.command) {
-                        case 'refresh':
-                            ussFileProvider.refreshElement(node);
-                            await editView.panel.webview.postMessage({
-                                attributes: node.attributes,
-                                name: node.fullPath
-                            });
-                            break;
-                        case 'ready': 
-                            await editView.panel.webview.postMessage({
-                                attributes: node.attributes,
-                                name: node.fullPath
-                            });
-                            break;
-                        case 'update-attributes':
-                            if (Object.keys(message.attrs).length > 0) {
-                                const attrs = message.attrs;
-                                if (node.attributes.owner !== attrs.owner) {
-                                    await Utilities.putUSSPayload(node.getSession(), node.fullPath, {
-                                        request: "chown",
-                                        owner: attrs.owner,
-                                        group: attrs.group,
-                                        links: "follow",
-                                        recursive: true
-                                    });
-                                    node.attributes.owner = attrs.owner;
-                                }
-                                if (!isNaN(attrs.group) && !isNaN(parseFloat(attrs.group))) {
-                                    const gid = parseInt(attrs.group);
-                                    if (node.attributes.gid !== gid) {
-                                        await Utilities.putUSSPayload(node.getSession(), node.fullPath, {
-                                            request: "chown",
-                                            owner: attrs.owner,
-                                            group: attrs.gid,
-                                            links: "follow",
-                                            recursive: true
-                                        });
-                                        node.attributes.gid = gid;
-                                    }
-                                } else if (node.attributes.group !== attrs.group) {
-                                    await Utilities.putUSSPayload(node.getSession(), node.fullPath, {
-                                        request: "chown",
-                                        owner: attrs.owner,
-                                        group: attrs.group,
-                                        links: "follow",
-                                        recursive: true
-                                    });
-                                    node.attributes.group = attrs.group;
-                                }
-                                if (node.attributes.perms !== attrs.perms) {
-                                    const permsAsOctal = permStringToOctal(attrs.perms);
-                                    await Utilities.putUSSPayload(node.getSession(), node.fullPath, {
-                                        request: "chmod",
-                                        mode: permsAsOctal,
-                                        links: "follow",
-                                        recursive: true 
-                                    });
-                                    node.attributes.perms = attrs.perms;
-                                }
-                                await editView.panel.webview.postMessage({
-                                    updated: true
-                                });
-                                await Gui.infoMessage(`Updated file attributes for ${node.fullPath}`);
-                            }
-                            break;
-                        default: break;
-                    }
-                }
-            );
-        })
+        vscode.commands.registerCommand("zowe.uss.editAttributes", (node: IZoweUSSTreeNode) =>
+            ussActions.editAttributes(context, ussFileProvider, node)
+        )
     );
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.uss.saveSearch", (node: IZoweUSSTreeNode): void => ussFileProvider.saveSearch(node))
